refactor: reorganize project structure
https://github.com/Attacktive/troubleshooter-editor-front-end-svelte/issues/14
