+++ src/routes/+layout.svelte
@@ -1,7 +1,7 @@
 <script>
 	import "../app.css";
-	import Header from "./Header.svelte";
-	import Footer from "./Footer.svelte";
+	import Header from "../components/Header.svelte";
+	import Footer from "../components/Footer.svelte";
 </script>
 
 <div class="p-3">
